travel_advisor/requirements.txt
a2a-sdk
langchain
langchain-core
langchain-openai
httpx
uvicorn